Break lb-radio module into several modules, one per class.
